+++ common/site/userprofileadmin.py
@@ -1,7 +1,3 @@
-from PIL import Image
-from PIL import ImageDraw
-from io import BytesIO
-
 from django.conf import settings
 from django.contrib import admin
 from django.contrib.auth.models import User
@@ -15,6 +11,7 @@ from django.utils.translation import gettext_lazy as _
 
 from common.forms.userprofileform import UserProfileForm
 from common.models import UserProfile
+from common.utils.resize_image import resize_image
 from common.utils.admfilters import ByDepartmentFilter
 from common.utils.chat_link import get_chat_link
 from common.utils.helpers import add_chat_context
@@ -254,30 +251,12 @@ class UserProfileAdmin(admin.ModelAdmin):
 
 # -- Custom Methods -- #
 
-
 def resize_avatar(obj) -> None:
     """
     Resize uploaded avatar image to a maximum of 200x200 pixels.
     """
     if obj.avatar:
-        img = Image.open(obj.avatar)
-        img.thumbnail((200, 200), Image.Resampling.LANCZOS)
-
-        # Convert to RGB if necessary (for PNG with transparency)
-        if img.mode in ('RGBA', 'LA', 'P'):
-            img = img.convert('RGB')
-
-        # Create circular mask
-        size = img.size
-        mask = Image.new('L', size, 0)
-        draw = ImageDraw.Draw(mask)
-        draw.ellipse([0, 0, size[0], size[1]], fill=255)
-        img.putalpha(mask)
-
-        # Save to BytesIO
-        resized_image = BytesIO()
-        img.save(resized_image, format='PNG', quality=85)
-        resized_image.seek(0)
+        resized_image = resize_image(obj.avatar)
 
         # Create a new File object
         obj.avatar.file = resized_image