+++ CHANGELOG.md
@@ -10,9 +10,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Fixed
+
+- Fix weight field change handling
+
 ### Improved
 
 - Refactor contact admin field layout for improved readability
+
+### Changed
+
+- Moved the resize image functionality to common utils
 
 ## [2.3.1] - 2026-04-26